+++ docs/source/changelog.rst
@@ -8,6 +8,8 @@ Changelog
 
 - Deprecates ``auth_local_webserver`` argument in favor of
   ``use_local_webserver`` argument (:issue:`20`).
+- Adds :func:`pydata_google_auth.save_user_credentials` function to get user
+  credentials and then save them to a specified JSON path. (:issue:`22`)
 
 .. _changelog-0.1.3: